robot-register/requirements.txt
paho-mqtt==2.0.0
PyYAML==6.0.1
